GIT: rename monte carlo files
