+++ rocketpy/monte_carlo/mc_aero_surfaces.py
@@ -1,21 +1,11 @@
 from typing import Any, List, Tuple, Union
 
-from pydantic import (
-    BaseModel,
-    Field,
-    FilePath,
-    StrictFloat,
-    StrictInt,
-    StrictStr,
-    ValidationError,
-    root_validator,
-    validator,
-)
+from pydantic import BaseModel, Field, FilePath, StrictInt, StrictStr, root_validator
 
-from .AeroSurfaces import EllipticalFins, NoseCone, Tail, TrapezoidalFins
+from ..AeroSurfaces import EllipticalFins, NoseCone, Tail, TrapezoidalFins
 
 
-class NoseConeData(BaseModel):
+class McNoseCone(BaseModel):
     nosecone: NoseCone = Field(..., repr=False)
     length: Any = 0
     kind: List[Union[StrictStr, None]] = []
@@ -32,15 +22,15 @@ class NoseConeData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
         validate_fields = [
@@ -64,7 +54,7 @@ class NoseConeData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
@@ -78,7 +68,7 @@ class NoseConeData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
@@ -103,7 +93,7 @@ class NoseConeData(BaseModel):
         return values
 
 
-class TrapezoidalFinsData(BaseModel):
+class McTrapezoidalFins(BaseModel):
     trapezoidalFins: TrapezoidalFins = Field(..., repr=False)
     n: List[StrictInt] = []
     rootChord: Any = 0
@@ -125,15 +115,15 @@ class TrapezoidalFinsData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
         validate_fields = [
@@ -162,7 +152,7 @@ class TrapezoidalFinsData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
@@ -180,7 +170,7 @@ class TrapezoidalFinsData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
@@ -205,7 +195,7 @@ class TrapezoidalFinsData(BaseModel):
         return values
 
 
-class EllipticalFinsData(BaseModel):
+class McEllipticalFins(BaseModel):
     ellipticalFins: EllipticalFins = Field(..., repr=False)
     n: Any = 0
     rootChord: Any = 0
@@ -224,15 +214,15 @@ class EllipticalFinsData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
         validate_fields = [
@@ -258,7 +248,7 @@ class EllipticalFinsData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
@@ -276,7 +266,7 @@ class EllipticalFinsData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
@@ -301,7 +291,7 @@ class EllipticalFinsData(BaseModel):
         return values
 
 
-class TailData(BaseModel):
+class McTail(BaseModel):
     tail: Tail = Field(..., repr=False)
     topRadius: Any = 0
     bottomRadius: Any = 0
@@ -318,15 +308,15 @@ class TailData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
         validate_fields = [
@@ -350,7 +340,7 @@ class TailData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
@@ -364,7 +354,7 @@ class TailData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
