+++ rocketpy/monte_carlo/mc_environment.py
@@ -1,21 +1,11 @@
 from typing import Any, List, Tuple, Union
 
-from pydantic import (
-    BaseModel,
-    Field,
-    FilePath,
-    StrictFloat,
-    StrictInt,
-    StrictStr,
-    ValidationError,
-    validator,
-    root_validator,
-)
+from pydantic import BaseModel, Field, StrictStr, root_validator
 
-from .Environment import Environment
+from ..Environment import Environment
 
 
-class EnvironmentData(BaseModel):
+class McEnvironment(BaseModel):
     """Class that holds that about an environment to be used in the Dispersion class.
 
     Parameters
@@ -30,8 +20,8 @@ class EnvironmentData(BaseModel):
         object, and the second item is the value passed. This value represents the
         standard deviation to be used in the Dispersion. If a tuple is passed, the first
         item must be the standard deviation and the second must be a string containing the
-        name of the distribuition function to be used in the dispersion. If no string is
-        passed, a normal distribuition is used.
+        name of the distribution function to be used in the dispersion. If no string is
+        passed, a normal distribution is used.
     date : list, optional
         Array of length 4, stating (year, month, day, hour (UTC)) of rocket launch.
         Must be given if a Forecast, Reanalysis or Ensemble, will be set as an
@@ -47,8 +37,8 @@ class EnvironmentData(BaseModel):
         object, and the second item is the value passed. This value represents the
         standard deviation to be used in the Dispersion. If a tuple is passed, the first
         item must be the standard deviation and the second must be a string containing the
-        name of the distribuition function to be used in the dispersion. If no string is
-        passed, a normal distribuition is used.
+        name of the distribution function to be used in the dispersion. If no string is
+        passed, a normal distribution is used.
     gravity : int, float, tuple, optional
         Surface gravitational acceleration. Positive values point the acceleration down.
         Default value is 9.80665. Should be given as an int or
@@ -57,8 +47,8 @@ class EnvironmentData(BaseModel):
         object, and the second item is the value passed. This value represents the
         standard deviation to be used in the Dispersion. If a tuple is passed, the first
         item must be the standard deviation and the second must be a string containing the
-        name of the distribuition function to be used in the dispersion. If no string is
-        passed, a normal distribuition is used.
+        name of the distribution function to be used in the dispersion. If no string is
+        passed, a normal distribution is used.
     latitude : int, float, tuple, optional
         Latitude in degrees (ranging from -90 to 90) of rocket launch location. Must be
         given if a Forecast, Reanalysis or Ensemble will be used as an atmospheric model
@@ -68,10 +58,10 @@ class EnvironmentData(BaseModel):
         object, and the second item is the value passed. This value represents the
         standard deviation to be used in the Dispersion. If a tuple is passed, the first
         item must be the standard deviation and the second must be a string containing the
-        name of the distribuition function to be used in the dispersion. If no string is
-        passed, a normal distribuition is used.
+        name of the distribution function to be used in the dispersion. If no string is
+        passed, a normal distribution is used.
     longitude : int, float, tuple, optional
-        Longitude in degrees (ranging from -180 to 360) of rocket launch location.
+        Longitude in degrees (ranging from 180 to 360) of rocket launch location.
         Must be given if a Forecast, Reanalysis or Ensemble will be used as an atmospheric
         model or if Open-Elevation will be used to compute elevation. Should be given as an int or
         float, or a tuple. If int or float, the value passed will be used to construct a
@@ -79,8 +69,8 @@ class EnvironmentData(BaseModel):
         object, and the second item is the value passed. This value represents the
         standard deviation to be used in the Dispersion. If a tuple is passed, the first
         item must be the standard deviation and the second must be a string containing the
-        name of the distribuition function to be used in the dispersion. If no string is
-        passed, a normal distribuition is used.
+        name of the distribution function to be used in the dispersion. If no string is
+        passed, a normal distribution is used.
     datum : list, optional
         The desired reference ellipsoidal model, the following options are available:
         "SAD69", "WGS84", "NAD83", and "SIRGAS2000". The default is "SIRGAS2000", then
@@ -111,15 +101,15 @@ class EnvironmentData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
         validate_fields = [
@@ -145,7 +135,7 @@ class EnvironmentData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
@@ -163,7 +153,7 @@ class EnvironmentData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
