+++ rocketpy/monte_carlo/mc_rocket.py
@@ -1,29 +1,15 @@
 from typing import Any, List, Tuple, Union
 
-from pydantic import (
-    BaseModel,
-    Field,
-    FilePath,
-    PrivateAttr,
-    StrictStr,
-    root_validator,
-    validator,
-)
+from pydantic import BaseModel, Field, FilePath, PrivateAttr, root_validator
 
-from .ParachuteData import ParachuteData
-
-from .AeroSurfacesData import (
-    EllipticalFinsData,
-    NoseConeData,
-    TailData,
-    TrapezoidalFinsData,
-)
-from .Rocket import Rocket
+from ..Rocket import Rocket
+from .mc_aero_surfaces import McEllipticalFins, McNoseCone, McTail, McTrapezoidalFins
+from .mc_parachute import McParachute
 
 
 # TODO: make a special validator for power on and off factor since they need to have the nominal
 # value inputted
-class RocketData(BaseModel):
+class McRocket(BaseModel):
     rocket: Rocket = Field(..., repr=False)
     radius: Any = 0
     mass: Any = 0
@@ -76,15 +62,15 @@ class RocketData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
         validate_fields = [
@@ -111,7 +97,7 @@ class RocketData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
@@ -125,7 +111,7 @@ class RocketData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
@@ -150,33 +136,33 @@ class RocketData(BaseModel):
         return values
 
     def addNose(self, nosecone):
-        # checks if input is a noseconeData type
-        if not isinstance(nosecone, NoseConeData):
-            raise TypeError("nosecone must be of NoseConeData type")
+        # checks if input is a McNoseCone type
+        if not isinstance(nosecone, McNoseCone):
+            raise TypeError("nosecone must be of McNoseCone type")
         return self.nosecones.append(nosecone)
 
-    def addTrapezoidalFins(self, trapezoildalFins):
-        # checks if input is a noseconeData type
-        if not isinstance(trapezoildalFins, TrapezoidalFinsData):
-            raise TypeError("trapezoildalFins must be of NoseConeData type")
-        return self.fins.append(trapezoildalFins)
+    def addTrapezoidalFins(self, trapezoidalFins):
+        # checks if input is a McNoseCone type
+        if not isinstance(trapezoidalFins, McTrapezoidalFins):
+            raise TypeError("trapezoidalFins must be of McNoseCone type")
+        return self.fins.append(trapezoidalFins)
 
     def addEllipticalFins(self, ellipticalFins):
-        # checks if input is a noseconeData type
-        if not isinstance(ellipticalFins, EllipticalFinsData):
-            raise TypeError("ellipticalFins must be of NoseConeData type")
+        # checks if input is a McNoseCone type
+        if not isinstance(ellipticalFins, McEllipticalFins):
+            raise TypeError("ellipticalFins must be of McNoseCone type")
         return self.fins.append(ellipticalFins)
 
     def addTail(self, tail):
-        # checks if input is a noseconeData type
-        if not isinstance(tail, TailData):
-            raise TypeError("nosecone must be of NoseConeData type")
+        # checks if input is a McNoseCone type
+        if not isinstance(tail, McTail):
+            raise TypeError("nosecone must be of McNoseCone type")
         return self.tails.append(tail)
 
     def addParachute(self, parachute):
-        # checks if input is a noseconeData type
-        if not isinstance(parachute, ParachuteData):
-            raise TypeError("parachute must be of ParachuteData type")
+        # checks if input is a McNoseCone type
+        if not isinstance(parachute, McParachute):
+            raise TypeError("parachute must be of McParachute type")
         return self.parachutes.append(parachute)
 
     def addRailButtons(self, position1, position2, angle):
