+++ rocketpy/monte_carlo/mc_solid_motor.py
@@ -1,37 +1,25 @@
-from typing import Any, Callable, List, Tuple, Union
+from typing import Any, List, Tuple, Union
 
-from pydantic import (
-    BaseModel,
-    Field,
-    FilePath,
-    PrivateAttr,
-    StrictFloat,
-    StrictInt,
-    StrictStr,
-    root_validator,
-    validator,
-)
+from pydantic import BaseModel, Field, FilePath, StrictFloat, StrictInt, root_validator
 
-from .Parachute import Parachute
+from ..Motor import SolidMotor
 
 
-class ParachuteData(BaseModel):
-    parachute: Parachute = Field(..., repr=False)
-    CdS: Any = 0
-    trigger: List[Union[Callable, None]] = []
-    samplingRate: Any = 0
-    lag: Any = 0
-    name: List[Union[StrictStr, None]] = []
-    noise: List[
-        Union[
-            Tuple[
-                Union[StrictInt, StrictFloat],
-                Union[StrictInt, StrictFloat],
-                Union[StrictInt, StrictFloat],
-            ],
-            None,
-        ]
-    ] = []
+class McSolidMotor(BaseModel):
+    solidMotor: SolidMotor = Field(..., repr=False)
+    thrustSource: List[Union[FilePath, None]] = []
+    burnOutTime: Any = 0
+    grainsCenterOfMassPosition: Any = 0
+    grainNumber: List[Union[Union[StrictInt, StrictFloat], None]] = []
+    grainDensity: Any = 0
+    grainOuterRadius: Any = 0
+    grainInitialInnerRadius: Any = 0
+    grainInitialHeight: Any = 0
+    grainSeparation: Any = 0
+    totalImpulse: Any = 0
+    nozzleRadius: Any = 0
+    nozzlePosition: Any = 0
+    throatRadius: Any = 0
 
     class Config:
         arbitrary_types_allowed = True
@@ -41,18 +29,32 @@ class ParachuteData(BaseModel):
         """Validates inputs that can be either tuples, lists, ints or floats and
         saves them in the format (nom_val,std) or (nom_val,std,'dist_func').
         Lists are saved as they are inputted.
-        Inputs can be given as floats or ints, refering to the standard deviation.
+        Inputs can be given as floats or ints, referring to the standard deviation.
         In this case, the nominal value of that attribute will come from the rocket
-        object passed. If the distribuition function whants to be specified, then a
+        object passed. If the distribution function needs to be specified, then a
         tuple with the standard deviation as the first item, and the string containing
-        the name a numpy.random distribuition function can be passed.
+        the name a numpy.random distribution function can be passed.
         If a tuple with a nominal value and a standard deviation is passed, then it
-        will take priority over the rocket object attriubute's value. A third item
-        can also be added to the tuple specifying the distribuition function"""
-        # TODO: add a way to check if the strings refering to the distribuition func
+        will take priority over the rocket object attribute's value. A third item
+        can also be added to the tuple specifying the distribution function"""
+        # TODO: add a way to check if the strings referring to the distribution func
         # are valid names for numpy.random functions
 
-        validate_fields = ["CdS", "trigger", "samplingRate", "lag", "name", "noise"]
+        validate_fields = [
+            "thrustSource",
+            "burnOutTime",
+            "grainsCenterOfMassPosition",
+            "grainNumber",
+            "grainDensity",
+            "grainOuterRadius",
+            "grainInitialInnerRadius",
+            "grainInitialHeight",
+            "grainSeparation",
+            "nozzleRadius",
+            "nozzlePosition",
+            "throatRadius",
+            "totalImpulse",
+        ]
         for field in validate_fields:
             v = values[field]
             # checks if tuple
@@ -66,14 +68,14 @@ class ParachuteData(BaseModel):
                     # checks if second value is either string or int/float
                     assert isinstance(
                         v[1], (int, float, str)
-                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tSecond item of tuple must be either an int, float or string \n  If the first value refers to the nominal value of {field}, then the second item's value should be the desired standard deviation \n  If the first value is the standard deviation, then the second item's value should be a string containing a name of a numpy.random distribution function"
                     # if second item is not str, then (nom_val, std)
                     if not isinstance(v[1], str):
                         values[field] = v
                     # if second item is str, then (nom_val, std, str)
                     else:
                         values[field] = (
-                            getattr(values["parachute"], field),
+                            getattr(values["solidMotor"], field),
                             v[0],
                             v[1],
                         )
@@ -84,13 +86,13 @@ class ParachuteData(BaseModel):
                     ), f"\nField '{field}': \n\tSecond item of tuple must be either an int or float \n  The second item should be the standard deviation to be used in the simulation"
                     assert isinstance(
                         v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribuition function"
+                    ), f"\nField '{field}': \n\tThird item of tuple must be a string \n  The string should contain the name of a valid numpy.random distribution function"
                     values[field] = v
             elif isinstance(v, list):
                 # checks if input list is empty, meaning nothing was inputted
                 # and values should be gotten from class
                 if len(v) == 0:
-                    values[field] = [getattr(values["parachute"], field)]
+                    values[field] = [getattr(values["solidMotor"], field)]
                 else:
                     # guarantee all values are valid (ints or floats)
                     assert all(
@@ -101,7 +103,7 @@ class ParachuteData(BaseModel):
             elif isinstance(v, (int, float)):
                 # not list or tuple, must be an int or float
                 # get attr and returns (nom_value, std)
-                values[field] = (getattr(values["parachute"], field), v)
+                values[field] = (getattr(values["solidMotor"], field), v)
             else:
                 raise ValueError(
                     f"\nField '{field}': \n\tMust be either a tuple, list, int or float"
